functions/requirements.txt
firebase-admin
firebase-functions
Flask
google-cloud-firestore
google-generativeai
